BP 5.10 — move BenefitRulePredicate types into engine domain
Relocates BenefitRulePredicate, BenefitMemberGender, and
BenefitRuleEvaluationContext from BenefitPlanService.Models.Benefits to
CloudHealthOffice.BenefitEngine.Domain so the engine can reference
them without a circular dependency on benefit-plan-service. The shape
and JSON wire format are unchanged; consumers in benefit-plan-service
gain a `using CloudHealthOffice.BenefitEngine.Domain;` import.

Preparation for the BP 5.10 BenefitRuleGate that will run inside the
calculation engine.
