+++ src/services/benefit-plan-service/BenefitPlanService.Tests/Models/Benefits/BenefitRulePredicateTests.cs
@@ -1,4 +1,5 @@
 using BenefitPlanService.Models.Benefits;
+using CloudHealthOffice.BenefitEngine.Domain;
 
 namespace BenefitPlanService.Tests.Models.Benefits;
