+++ src/services/benefit-plan-service/Models/BenefitPlanAdapterResponse.cs
@@ -1,5 +1,6 @@
 using System.Text.Json.Serialization;
 using BenefitPlanService.Models.Benefits;
+using CloudHealthOffice.BenefitEngine.Domain;
 
 namespace BenefitPlanService.Models;
